% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchSize.R
\name{batchSize}
\alias{batchSize}
\title{Batch size (truncation point) selection}
\usage{
batchSize(x, method = "bm", g = NULL)
}
\arguments{
\item{x}{A matrix or data frame of Markov chain output. Number of rows is the Monte
Carlo sample size.}

\item{method}{Any of `bm`,`obm`,`bartlett`,`tukey`. `bm` represents batch
means estimator, `obm` represents the overlapping batch means estimator,
and `bartlett` and `tukey` represent the modified-Bartlett window and
the Tukey-Hanning windows for the spectral variance estimators.}

\item{g}{A function that represents features of interest. g is applied to each row of x and
thus g should take a vector input only. If g is NULL, g is set to be identity, which
is estimation of the mean of the target density.}
}
\value{
A value of the optimal batch size is returned.
}
\description{
Function returns the optimal batch size (or truncation point) for a given chain and method.
}
\details{
Final update of batchSize using thresholding on aR coefficients and option to use only the tail of the chain for acf
calculation.
}
\examples{
library(mAr)
p <- 3
n <- 1e3
omega <- 5*diag(1,p)
## Making correlation matrix var(1) model
set.seed(100)
foo <- matrix(rnorm(p^2), nrow = p)
foo <- foo \%*\% t(foo)
phi <- foo / (max(eigen(foo)$values) + 1)
out <- as.matrix(mAr.sim(rep(0,p), phi, omega, N = n))
batchSize(out)
batchSize(out, method = "obm")
batchSize(out, method = "bartlett")

## Bivariate Normal with mean (mu1, mu2) and covariance sigma
n <- 1e3
mu = c(2, 50)
sigma = matrix(c(1, 0.5, 0.5, 1), nrow = 2)
X = BVN_Gibbs(n, mu, sigma)
batchSize(X)
batchSize(X, method = "obm")
batchSize(X, method = "bartlett")

}
\references{
Liu, Y., Vats, D., and Flegal, J. M. Batch size selection for variance estimators in MCMC, arXiv
preprint arXiv:1804.05975 (2019).
}
\seealso{
\code{\link{mcse.multi}}, which calls on batchSize. \code{\link{mcse}}, which calls on batchSize.
}
