% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVG.R
\name{BVN_Gibbs}
\alias{BVN_Gibbs}
\title{MCMC on a bivariate normal distribution}
\usage{
BVN_Gibbs(n, mu, sigma)
}
\arguments{
\item{n}{Sample size of the Markov chain}

\item{mu}{A 2 dimensional vectoe. Mean of the target normal distribution}

\item{sigma}{2 x 2 symmetric positive semi-definite matrix. The covariance matrix of the target normal distribution.}
}
\value{
A Markov chain sampled from the taret normal distribution.
}
\description{
Function returns a Markov chain sampled using Gibbs sampling from a bivariate normal target
}
\examples{
n <- 1e3
mu = c(2, 50)
sigma = matrix(c(1, 0.5, 0.5, 1), nrow = 2)
X = BVN_Gibbs(n, mu, sigma)

}
