% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcse_multi.R
\name{is.mcmcse}
\alias{is.mcmcse}
\title{Check if the class of the object is mcmcse}
\usage{
is.mcmcse(x)
}
\arguments{
\item{x}{The object that is checked to belong to the class mcmcse}
}
\value{
Boolean variable indicating if the input belongs to the class mcmcse
}
\description{
Check if the class of the object is mcmcse
}
\examples{

## Bivariate Normal with mean (mu1, mu2) and covariance sigma
n <- 1e3
mu <- c(2, 50)
sigma <- matrix(c(1, 0.5, 0.5, 1), nrow = 2)

out <- BVN_Gibbs(n, mu, sigma)
is.mcmcse(mcse.multi(out))

}
