% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confRegion.R
\name{confRegion}
\alias{confRegion}
\title{Confidence regions (ellipses) for Monte Carlo estimates}
\usage{
confRegion(mcse.obj, which = c(1,2), level = .95)
}
\arguments{
\item{mcse.obj}{The list returned by the mcse.multi or mcse.initseq command.}

\item{which}{Integer vector of length 2 indicating the component for which to make the confidence
ellipse. Chooses the first two by #'  default.}

\item{level}{confidence level for the ellipse.}
}
\value{
Returns a matrix of x and y coordinates for the ellipse. Use plot function on the matrix
to plot the ellipse.
}
\description{
Constructs confidence regions (ellipses) from the Markov chain output for the features of interest
Function uses the ellipse package.
}
\examples{
library(mAr)
p <- 3
n <- 1e3
omega <- 5*diag(1,p)
## Making correlation matrix var(1) model
set.seed(100)
foo <- matrix(rnorm(p^2), nrow = p)
foo <- foo \%*\% t(foo)
phi <- foo / (max(eigen(foo)$values) + 1)
out <- as.matrix(mAr.sim(rep(0,p), phi, omega, N = n))
mcerror <- mcse.multi(out, blather = TRUE)
## Plotting the ellipse
plot(confRegion(mcerror), type = 'l')

## Bivariate Normal with mean (mu1, mu2) and covariance sigma
n <- 1e3
mu = c(2, 50)
sigma = matrix(c(1, 0.5, 0.5, 1), nrow = 2)
X = multivariate_Gibbs_normal(n, mu, sigma)
mcerror <- mcse.multi(X, blather = TRUE)
## Plotting the ellipse
plot(confRegion(mcerror), type = 'l')

}
