% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_test.R
\name{qqTest}
\alias{qqTest}
\title{QQplot for Markov chains}
\usage{
qqTest(mcse.obj)
}
\arguments{
\item{mcse.obj}{the list returned by the `mcse.multi` or `mcse.initseq` command}
}
\description{
QQplot for Markov chains using an estimate of the Markov Chain CLT covariance matrix.
}
\examples{
library(mAr)
p <- 35
n <- 1e4
omega <- 5*diag(1,p)
## Making correlation matrix var(1) model
set.seed(100)
foo <- matrix(rnorm(p^2), nrow = p)
foo <- foo \%*\% t(foo)
phi <- foo / (max(eigen(foo)$values) + 1)
out <- as.matrix(mAr.sim(rep(0,p), phi, omega, N = n))
mcse.bm <- mcse.multi(x = out)
qqTest(mcse.bm)
mcse.isadj <- mcse.initseq(x = out, adjust = TRUE)
qqTest(mcse.isadj)

## Bivariate Normal with mean (mu1, mu2) and covariance sigma
n <- 1e3
mu = c(2, 50)
sigma = matrix(c(1, 0.5, 0.5, 1), nrow = 2)
X = BVN_Gibbs(n, mu, sigma)
mcse.bm <- mcse.multi(x = X)
qqTest(mcse.bm)
mcse.isadj <- mcse.initseq(x = X, adjust = TRUE)
qqTest(mcse.isadj)

}
